// src/lib/sanity/schema-types.ts
export type PortableText = Array<
  | {
      children?: Array<{
        marks?: Array<string>
        text?: string
        _type: 'span'
        _key: string
      }>
      style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote'
      listItem?: 'bullet' | 'number'
      markDefs?: Array<
        | ({
            _key: string
          } & LinkProduct)
        | ({
            _key: string
          } & LinkEmail)
        | ({
            _key: string
          } & LinkInternal)
        | ({
            _key: string
          } & LinkExternal)
      >
      level?: number
      _type: 'block'
      _key: string
    }
  | ({
      _key: string
    } & Accordion)
  | ({
      _key: string
    } & Callout)
  | ({
      _key: string
    } & Grid)
  | ({
      _key: string
    } & Images)
  | ({
      _key: string
    } & ImageWithProductHotspots)
  | ({
      _key: string
    } & Instagram)
  | ({
      _key: string
    } & Products)
>
// ...
export type ProxyString = string
// ...
export type Products = {
  _type: 'products'
  title?: string
  layout?: 'card' | 'pill'
  limit?: number
}

export type PriceRange = {
  _type: 'priceRange'
  minVariantPrice?: number
  maxVariantPrice?: number
}

export type PlaceholderString = string

export type Option = {
  _type: 'option'
  name?: string
  values?: Array<string>
}
// ...
export type Instagram = {
  _type: 'instagram'
  url?: string
}

export type ImageWithProductHotspots = {
  _type: 'imageWithProductHotspots'
  image?: {
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    media?: unknown
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    _type: 'image'
  }
  showHotspots?: boolean
  productHotspots?: ProductHotspots
}
// ...
export type Images = {
  _type: 'images'
  imageFeatures?: Array<
    {
      _key: string
    } & ImageFeature
  >
  fullWidth?: boolean
  verticalAlign?: 'top' | 'center' | 'bottom'
}

export type ImageCallToAction = {
  _type: 'imageCallToAction'
  title?: string
  link?: Array<
    | ({
        _key: string
      } & LinkInternal)
    | ({
        _key: string
      } & LinkExternal)
  >
}

export type HeadingBlock = {
  _type: 'headingBlock'
  heading?: string
  level?: 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6'
  alignment?: 'left' | 'center' | 'right'
}

export type Grid = {
  _type: 'grid'
  items?: Array<
    {
      _key: string
    } & GridItem
  >
}
// ...
export type Callout = {
  _type: 'callout'
  text?: string
  link?: Array<
    | ({
        _key: string
      } & LinkInternal)
    | ({
        _key: string
      } & LinkExternal)
  >
}

export type Accordion = {
  _type: 'accordion'
  groups?: Array<
    {
      _key: string
    } & AccordionGroup
  >
}

export type AccordionGroup = {
  _type: 'accordionGroup'
  title?: string
  body?: PortableTextSimple
}

export type LinkProduct = {
  _type: 'linkProduct'
  productWithVariant?: ProductWithVariant
  linkAction?: 'link' | 'addToCart' | 'buyNow'
  quantity?: number
}

export type ProductWithVariant = {
  _type: 'productWithVariant'
  product?: {
    _ref: string
    _type: 'reference'
    _weak?: boolean
    [internalGroqTypeReferenceTo]?: 'product'
  }
  variant?: {
    _ref: string
    _type: 'reference'
    _weak?: boolean
    [internalGroqTypeReferenceTo]?: 'productVariant'
  }
}

export type LinkInternal = {
  _type: 'linkInternal'
  label?: string
  reference?:
    | {
        _ref: string
        _type: 'reference'
        _weak?: boolean
        [internalGroqTypeReferenceTo]?: 'collection'
      }
    | {
        _ref: string
        _type: 'reference'
        _weak?: boolean
        [internalGroqTypeReferenceTo]?: 'home'
      }
    | {
        _ref: string
        _type: 'reference'
        _weak?: boolean
        [internalGroqTypeReferenceTo]?: 'page'
      }
    | {
        _ref: string
        _type: 'reference'
        _weak?: boolean
        [internalGroqTypeReferenceTo]?: 'product'
      }
}
// ...
export type LinkExternal = {
  _type: 'linkExternal'
  url?: string
  newWindow?: boolean
}

export type LinkEmail = {
  _type: 'linkEmail'
  email?: string
}
// ...
export type Product = {
  _id: string
  _type: 'product'
  _createdAt: string
  _updatedAt: string
  _rev: string
  hidden?: string
  titleProxy?: ProxyString
  slugProxy?: ProxyString
  colorTheme?: {
    _ref: string
    _type: 'reference'
    _weak?: boolean
    [internalGroqTypeReferenceTo]?: 'colorTheme'
  }
  body?: PortableText
  store?: ShopifyProduct
  seo?: Seo
}

export type ShopifyProduct = {
  _type: 'shopifyProduct'
  createdAt?: string
  updatedAt?: string
  status?: 'active' | 'archived' | 'draft'
  isDeleted?: boolean
  title?: string
  id?: number
  gid?: string
  slug?: Slug
  descriptionHtml?: string
  productType?: string
  vendor?: string
  tags?: string
  priceRange?: PriceRange
  previewImageUrl?: string
  options?: Array<
    {
      _key: string
    } & Option
  >
  variants?: Array<{
    _ref: string
    _type: 'reference'
    _weak?: boolean
    _key: string
    [internalGroqTypeReferenceTo]?: 'productVariant'
  }>
}
// ...
export type Seo = {
  _type: 'seo'
  title?: string
  description?: string
  image?: {
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    media?: unknown
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    _type: 'image'
  }
}
// ...
export type SanityImageHotspot = {
  _type: 'sanity.imageHotspot'
  x?: number
  y?: number
  height?: number
  width?: number
}

export type SanityImageCrop = {
  _type: 'sanity.imageCrop'
  top?: number
  bottom?: number
  left?: number
  right?: number
}
// ...
export type Slug = {
  _type: 'slug'
  current?: string
  source?: string
}
// ...
export declare const internalGroqTypeReferenceTo: unique symbol